requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
moviepy==1.0.3
gTTS==2.5.1
pydub==0.25.1
